requirements.txt
Flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.0
joblib==1.3.2
python-docx==0.8.11
PyMuPDF==1.23.5
numpy==1.24.3
pandas==1.5.3
nltk==3.8.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
transformers==4.33.2
torch==2.0.1
werkzeug==2.3.7